+++ BitmovinAnalyticsCollector/Classes/Collector/analytics/BitmovinAnalyticsInternal.swift
@@ -140,6 +140,11 @@ extension BitmovinAnalyticsInternal: StateMachineDelegate {
         sendEventData(eventData: eventData)
     }
     
+    func stateMachineDidAudioChange(_ stateMachine: StateMachine) {
+        let eventData = createEventData(duration: 0)
+        sendEventData(eventData: eventData)
+    }
+    
     var currentTime: CMTime? {
         get {
             return self.adapter?.currentTime