+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+## v1.8.0
+
 ### Added
 
 - `subtitleEnabled` and `subtitleLanguage` in outgoing payload